refactor: standardize on the components.json `cn` alias
`src/lib/cn.ts` and `src/lib/utils.ts` were byte-equivalent implementations
of the same helper, differing only in export style. `components.json` points
its `utils` alias at the latter, so anything installed by `shadcn add` imports
the named export while everything hand-written imported the default one.

Keep the aliased file, drop the other. Prerequisite for extracting a block:
a distributed component has to import through the alias the CLI rewrites.
