+++ src/components/Icon.tsx
@@ -1,4 +1,4 @@
-import cn from '@/lib/cn'
+import { cn } from '@/lib/utils'
 import * as React from 'react'
 
 type SVGProps = Partial<React.SVGProps<SVGSVGElement>>
