+++ src/components/mdx/Codesandbox/Codesandbox.tsx
@@ -1,6 +1,6 @@
 import { Img } from '@/components/mdx/Img'
 
-import cn from '@/lib/cn'
+import { cn } from '@/lib/utils'
 import { ComponentProps } from 'react'
 
 export type CSB = {
