+++ src/components/mdx/Grid.tsx
@@ -1,4 +1,4 @@
-import cn from '@/lib/cn'
+import { cn } from '@/lib/utils'
 
 /**
  * @deprecated
