+++ src/components/mdx/Keypoints/Keypoints.tsx
@@ -1,4 +1,4 @@
-import cn from '@/lib/cn'
+import { cn } from '@/lib/utils'
 import { ComponentProps } from 'react'
 
 import { ul as Ul, li as Li } from '../index'
