+++ src/components/mdx/People/People.tsx
@@ -1,4 +1,4 @@
-import cn from '@/lib/cn'
+import { cn } from '@/lib/utils'
 import { initials } from '@/utils/text'
 import { Octokit } from '@octokit/core'
 import Image from 'next/image'
