+++ src/components/mdx/index.tsx
@@ -1,4 +1,4 @@
-import cn from '@/lib/cn'
+import { cn } from '@/lib/utils'
 import { ComponentProps } from 'react'
 
 const MARKDOWN_REGEX = /\.mdx?/
